Amends issues from branch merge
